+++ contract-server/src/main/java/org/seekay/contract/server/servet/RequestHandlerServlet.java
@@ -3,10 +3,10 @@ package org.seekay.contract.server.servet;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.seekay.contract.common.enrich.EnricherService;
+import org.seekay.contract.common.match.MatchingService;
+import org.seekay.contract.common.service.ContractService;
 import org.seekay.contract.model.domain.Contract;
 import org.seekay.contract.model.domain.ContractRequest;
-import org.seekay.contract.common.ApplicationContext;
-import org.seekay.contract.common.match.MatchingService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -14,6 +14,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static org.seekay.contract.common.ApplicationContext.*;
 import static org.seekay.contract.server.util.RequestReader.from;
 import static org.seekay.contract.server.util.ResponseWriter.to;
 
@@ -21,18 +22,20 @@ public class RequestHandlerServlet extends HttpServlet {
 
     private MatchingService matchingService;
     private EnricherService enricherService;
-
+    private ContractService contractService;
 
     @Override
     public void init() throws ServletException {
-        matchingService = ApplicationContext.matchingService();
-        enricherService = ApplicationContext.enricherService();
+        matchingService = matchingService();
+        enricherService = enricherService();
+        contractService = contractService();
     }
 
     @Override
     protected void doGet(HttpServletRequest httpRequest, HttpServletResponse httpsResponse) throws ServletException, IOException {
         ContractRequest contractRequest = from(httpRequest).toContractRequest();
         Contract contract = matchingService.matchGetRequest(contractRequest);
+        enableFilters(contract);
         contract = enricherService.enrichResponse(contract);
         writeResponse(httpsResponse, contract);
     }
@@ -70,5 +73,9 @@ public class RequestHandlerServlet extends HttpServlet {
         } else {
             to(httpResponse).notFound();
         }
+    }
+
+    private void enableFilters(Contract contract) {
+        contractService.enableFilters(contract.getFilters());
     }
 }